+++ framework/Furion/EventBus/ChannelEventPublisher.cs
@@ -56,4 +56,54 @@ internal sealed partial class ChannelEventPublisher : IEventPublisher
 
         return Task.CompletedTask;
     }
+
+    /// <summary>
+    /// 发布一条消息
+    /// </summary>
+    /// <param name="eventId">事件 Id</param>
+    /// <param name="payload">事件承载（携带）数据</param>
+    /// <param name="cancellationToken"> 取消任务 Token</param>
+    /// <returns></returns>
+    public async Task PublishAsync(string eventId, object payload = default, CancellationToken cancellationToken = default)
+    {
+        await PublishAsync(new ChannelEventSource(eventId, payload, cancellationToken));
+    }
+
+    /// <summary>
+    /// 发布一条消息
+    /// </summary>
+    /// <param name="eventId">事件 Id</param>
+    /// <param name="payload">事件承载（携带）数据</param>
+    /// <param name="cancellationToken"> 取消任务 Token</param>
+    /// <returns></returns>
+    public async Task PublishAsync(Enum eventId, object payload = default, CancellationToken cancellationToken = default)
+    {
+        await PublishAsync(new ChannelEventSource(eventId, payload, cancellationToken));
+    }
+
+    /// <summary>
+    /// 延迟发布一条消息
+    /// </summary>
+    /// <param name="eventId">事件 Id</param>
+    /// <param name="delay">延迟数（毫秒）</param>
+    /// <param name="payload">事件承载（携带）数据</param>
+    /// <param name="cancellationToken"> 取消任务 Token</param>
+    /// <returns><see cref="Task"/> 实例</returns>
+    public async Task PublishDelayAsync(string eventId, long delay, object payload = default, CancellationToken cancellationToken = default)
+    {
+        await PublishDelayAsync(new ChannelEventSource(eventId, payload, cancellationToken), delay);
+    }
+
+    /// <summary>
+    /// 延迟发布一条消息
+    /// </summary>
+    /// <param name="eventId">事件 Id</param>
+    /// <param name="delay">延迟数（毫秒）</param>
+    /// <param name="payload">事件承载（携带）数据</param>
+    /// <param name="cancellationToken"> 取消任务 Token</param>
+    /// <returns><see cref="Task"/> 实例</returns>
+    public async Task PublishDelayAsync(Enum eventId, long delay, object payload = default, CancellationToken cancellationToken = default)
+    {
+        await PublishDelayAsync(new ChannelEventSource(eventId, payload, cancellationToken), delay);
+    }
 }